Time: 88 ms (9.27%), Space: 7 MB (100.00%) - LeetHub

+++ 149-max-points-on-a-line/149-max-points-on-a-line.cpp
@@ -16,7 +16,7 @@ public:
                 else {
                 
                     for (int k=0; k<n; ++k) {
-                        if (static_cast<long>(p[k][1]-p[i][1]) * static_cast<long>(p[j][0]-p[i][0]) == static_cast<long>(p[j][1]-p[i][1]) * static_cast<long>(p[k][0]-p[i][0])) ++cnt;
+                        if (static_cast<long>(p[k][1]-p[i][1]) * (p[j][0]-p[i][0]) == static_cast<long>(p[j][1]-p[i][1]) * (p[k][0]-p[i][0])) ++cnt;
                     }
                 }
